requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.27
pydantic==2.6.3
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
pytest==7.4.4
httpx==0.27.0
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.6.1
asyncio==3.4.3
gunicorn==21.2.0